rename: 컴포넌트 확장자명 변경 js -> jsx
